develope(newFeature): Свёрстан Alert об ошибке загрузки данных
Выполнены следующие работы:
1) Добавлен Extensions файл с реализацией метода universalErrorAlert который отображает алерт об ошибке загрузки данных;
2) Обновлены вью контроллеры(StatisticsViewController, UserCardViewController, UserCollectionViewController) с добавлением метода universalErrorAlert для показа алерта об ошибке загрузки данных;
3) Обновлён NftSerivce, в методе getNft используется кастомная структура GetNftRequest.
