+++ FakeNFT/Scenes /Statistics/UsersList/StatisticsViewController.swift
@@ -19,7 +19,7 @@ enum StatisticsViewControllerLayout {
     static let filterButtonSizes: CGFloat = 42
 }
 
-final class StatisticsViewController: UIViewController, LoadingView, ErrorView {
+final class StatisticsViewController: UIViewController, LoadingView {
     
     // MARK: - Private Properties
     private let viewModel: StatisticsViewModel
@@ -143,8 +143,6 @@ final class StatisticsViewController: UIViewController, LoadingView, ErrorView {
         case .hideBoth:
             self.hideLoading()
             self.tableViewWithUsers.isHidden = true
-        default:
-            break
         }
     }
     
@@ -178,8 +176,6 @@ final class StatisticsViewController: UIViewController, LoadingView, ErrorView {
                         showNeedViewsInScreen(needToShowLoadingIndicator: .showViewsHideLoader)
                         self.showErrorAlert()
                     }
-                default:
-                    break
                 }
             }
             .store(in: &cancellables)
@@ -190,26 +186,10 @@ final class StatisticsViewController: UIViewController, LoadingView, ErrorView {
     }
     
     private func showErrorAlert() {
-        let retryActionText = NSLocalizedString("Statistics.errorAlert.retryAction.text", comment: "")
-        let errorModel = ErrorModel(
-            message: "",
-            actionText: retryActionText,
-            action: {[weak self] in
-                guard let self else { return }
-                self.viewModel.fetchUsers()
-            }
-        )
-        
-        let title = NSLocalizedString("Statistics.errorAlert.title", comment: "")
-        
-        let retryAction = UIAlertAction(title: errorModel.actionText, style: .default) {_ in
-            errorModel.action()
+        self.universalErrorAlert {[weak self] in
+            guard let self else { return }
+            self.viewModel.fetchUsers()
         }
-        
-        let cancelActionText = NSLocalizedString("Statistics.errorAlert.cancelAction.text", comment: "")
-        let cancelAction = UIAlertAction(title: cancelActionText, style: .cancel)
-        
-        self.showErrorAlertWithTwoButtons(titleOfAlert: title, firstAction: cancelAction, secondAction: retryAction)
     }
 }
